added the loose PF jet ID

+++ include/Higgs.hh
@@ -55,7 +55,7 @@ public:
   // get the likelihood electron ID
   float likelihoodRatio(int eleIndex, ElectronLikelihood &lh);
   // get the PFjet ID
-  bool isPFJetID(double eta, double nHFrac, double nEmFrac, int nConst, double chHFrac, double chMult, double chEmFrac, int WP);
+  bool isPFJetID(float eta, float nHFrac, float nEmFrac, int nConst, float chHFrac, float chMult, float chEmFrac, int WP);
 
   enum jetIdWP { none=0, loose=1, medium=2, tight=3 };
